requirements.txt
openai>=1.30.0
langgraph>=0.2.0
langchain-core>=0.2.0
opentelemetry-sdk>=1.24.0
opentelemetry-api>=1.24.0
rich>=13.7.0
python-dotenv>=1.0.0
